refactor: 테스트 픽스처 네이밍 수정
