+++ src/test/java/com/snackgame/server/game/snackgame/core/domain/SnackgameTest.kt
@@ -1,7 +1,7 @@
 package com.snackgame.server.game.snackgame.core.domain
 
 
-import com.snackgame.server.game.snackgame.fixture.TestFixture
+import com.snackgame.server.game.snackgame.fixture.BoardFixture
 import com.snackgame.server.member.fixture.MemberFixture.땡칠
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -19,7 +19,7 @@ class SnackgameTest {
 
     @Test
     fun 스낵을_제거하면_점수를_얻을_수_있을() {
-        val game = Snackgame(땡칠().id, TestFixture.TWO_BY_FOUR())
+        val game = Snackgame(땡칠().id, BoardFixture.TWO_BY_FOUR())
         val streak = Streak(
             arrayListOf(
                 Coordinate(0, 0),
@@ -32,7 +32,7 @@ class SnackgameTest {
 
     @Test
     fun 황금스낵를_제거해도_점수는_초기화되지_않는다() {
-        val game = Snackgame(땡칠().id, TestFixture.TWO_BY_TWO_WITH_GOLDEN_SNACK())
+        val game = Snackgame(땡칠().id, BoardFixture.TWO_BY_TWO_WITH_GOLDEN_SNACK())
         val streak = Streak(
             arrayListOf(
                 Coordinate(0, 0),
